Greatly simplify autogen.sh and fix configure skipping
There is really no reason to manually vet and run all the autotools.
That's what autoreconf is for. It has the added bonus that it will
descend to the glib subdirectory and rebuild the autotools there. The
handling of configure is also fixed as the previous --no-configure code
was completely broken.

This is basically the xorg autogen.sh, so we can be pretty confident
it'll work as advertised.