popt: Reset long argument handling before iterating to next argument
From the LFS mailing list looking into why check-cmd-options fails using
the bundled popt:

poptGetNextOption, after it invokes the callback for
--define-variable=a=b (the operation of which sets the longArg local
variable), does not reset longArg to NULL.  So on the next time through
its loop (after invoking the --define-variable callback, before invoking
the --atleast-pkgconfig-version callback), on line 389 of popt.c, sets
con->os->nextArg to longArg (the old "a=b" string).  What it should do
is set nextArg to nextCharArg (the next item on the command line).  This
happens if longArg is NULL.

http://linuxfromscratch.org/pipermail/lfs-dev/2011-September/065065.html

Freedesktop #9584