Make sure glib subdir is configure before dist
We want to distribute the glib subdir, but we want to avoid configuring
it for the default build case where the system glib is used. Override
the standard dist target so that we can reconfigure with
--with-internal-glib if necessary. I think the target should be
compatible with the original dist.

This is pretty hacky and may need to be reverted if it breaks dist.