Remove pkg-config.in script
A relic from the past, the pkg-config.in script exists from a time when
pkg-config was implemented as a shell script. This time is long since
gone and the script is far different than the C implementation. Find it
in git if you want to see how a shell script once did pkg-config.